+++ R/sf_as_ee.R
@@ -10,7 +10,6 @@
 #' will be monitored.
 #' @param assetId Character. Destination asset ID for the uploaded file. Ignore
 #' if \code{via} argument is "getInfo".
-#' @param check_ring_dir Logical. See \code{st_read} for details.
 #' @param proj Integer or character. Coordinate Reference System (CRS) for the
 #' EE object, defaults to "EPSG:4326" (x=longitude, y=latitude).
 #' @param geodesic Logical. Ignored if \code{x} is not a Polygon or LineString.
@@ -26,6 +25,9 @@
 #' If unspecified, defaults to TRUE.
 #' @param bucket Character. Name of the bucket (GCS) to save intermediate files
 #' (ignore if \code{via} is not defined as "gcs_to_asset").
+#' @param command_line_tool_path Character. Path to the Earth Engine command line
+#' tool. If NULL, rgee assumes is saved in the same path that your Python
+#' environment (ignore if \code{via} is not defined as "gcs_to_asset").
 #' @param overwrite A boolean argument which indicates indicating
 #' whether "filename" should be overwritten. By default TRUE.
 #' @param quiet Logical. Suppress info message.
@@ -55,12 +57,6 @@
 #' See \href{https://developers.google.com/earth-engine/importing}{Importing
 #' table data} documentation for more details.
 #'
-#' Earth Engine is strict on polygon ring directions (outer ring
-#' counter-clockwise, and the inner one clockwise). If `check_ring_dir` is TRUE,
-#' it checks every ring, and revert them if necessary, to counter clockwise for
-#' outer, and clockwise for inner (hole) ones. By default, this is FALSE because
-#' it is an expensive operation.
-#'
 #' @examples
 #' \dontrun{
 #' library(rgee)
@@ -80,7 +76,10 @@
 #'                    ncol = 2,
 #'                    byrow = TRUE) %>%
 #'   list() %>%
-#'   st_polygon()
+#'   st_polygon() %>%
+#'   st_sfc(crs="EPSG:4326") %>%
+#'   st_sf()
+#'
 #' holePoly <- sf_as_ee(x = toy_poly, evenOdd = FALSE)
 #'
 #' # Create an even-odd version of the polygon.
@@ -99,10 +98,11 @@
 #'   x = toy_poly,
 #'   overwrite = TRUE,
 #'   assetId = assetId,
-#'   via = 'getInfo_to_asset')
+#'   via = "getInfo_to_asset")
 #'
 #' # 3. Upload large geometries to EE asset
-#' ee_Initialize(gcs = TRUE)
+#' ee_Initialize("data.colec.fbf", gcs = TRUE)
+#'
 #' assetId <- sprintf("%s/%s", ee_get_assethome(), 'toy_poly_gcs')
 #' eex <- sf_as_ee(
 #'   x = toy_poly,
@@ -112,38 +112,38 @@
 #'   monitoring = FALSE,
 #'   via = 'gcs_to_asset'
 #' )
+#' ee_monitoring()
 #' }
 #' @export
 sf_as_ee <- function(x,
                      via = 'getInfo',
                      assetId = NULL,
                      bucket = NULL,
+                     command_line_tool_path = NULL,
                      overwrite = TRUE,
                      monitoring = TRUE,
-                     check_ring_dir = FALSE,
                      evenOdd = TRUE,
                      proj = 4326,
                      geodesic = NULL,
                      quiet = FALSE,
                      ...) {
 
-  # Read geometry
-  eex <- ee_st_read(
-    x = x,
-    proj = proj,
-    check_ring_dir = check_ring_dir,
-    quiet = TRUE
-  )
+  if (!requireNamespace("sf", quietly = TRUE)) {
+    stop("package sf required, please install it first")
+  }
+
+  if (!any(class(x) %in%  "sf")) {
+    stop("x needs to be an object of class sf")
+  }
 
   # geodesic is null?
   if (is.null(geodesic)) {
-    is_geodesic <- sf::st_is_longlat(eex)
+    is_geodesic <- sf::st_is_longlat(x)
   } else {
     is_geodesic <- geodesic
   }
 
-
-  if (is.na(sf::st_crs(eex)$epsg)) {
+  if (is.na(sf::st_crs(x)$epsg)) {
     stop(
       "The x EPSG needs to be defined, use sf::st_set_crs to",
       " set, replace or retrieve."
@@ -151,23 +151,23 @@ sf_as_ee <- function(x,
   }
 
   # Transform x according to proj argument
-  eex_proj <- sf::st_crs(proj)$epsg
-  eex <- sf::st_transform(eex, eex_proj)
-  eex_proj <- sprintf("EPSG:%s", sf::st_crs(eex)$epsg)
+  x_proj <- sf::st_crs(proj)$epsg
+  x <- sf::st_transform(x, x_proj)
+  x_proj <- sprintf("EPSG:%s", sf::st_crs(x)$epsg)
 
   if (via == "getInfo") {
     # sf to geojson
     ee_sf_to_fc(
-      x = eex,
-      proj = eex_proj,
+      x = x,
+      proj = x_proj,
       geodesic = is_geodesic,
       evenOdd = evenOdd
     )
   } else if (via == "getInfo_to_asset") {
     # sf to geojson
     sf_fc <- ee_sf_to_fc(
-      x = eex,
-      proj = eex_proj,
+      x = x,
+      proj = x_proj,
       geodesic = is_geodesic,
       evenOdd = evenOdd
     )
@@ -180,10 +180,13 @@ sf_as_ee <- function(x,
     if (is.null(assetId)) {
       stop('assetId was not defined')
     }
+
+    # Verify is the EE assets path is valid
     assetId <- ee_verify_filename(
       path_asset = assetId,
       strict = FALSE
     )
+
     # geojson to assset
     ee_task <- ee_table_to_asset(
       collection = sf_fc,
@@ -202,17 +205,31 @@ sf_as_ee <- function(x,
     if (is.null(bucket)) {
       stop("Cloud Storage bucket was not defined")
     }
+
+    if (is.null(command_line_tool_path)) {
+      command_line_tool_path <- dirname(Sys.getenv("EARTHENGINE_PYTHON"))
+    }
+
     shp_dir <- sprintf("%s.shp", tempfile())
-    geozip_dir <- ee_utils_shp_to_zip(eex, shp_dir)
+
+    # From sf to .shp
+    message("1. Converting sf object to shapefile ... saving in /tmp")
+    geozip_dir <- ee_utils_shp_to_zip(x, shp_dir)
+
+    # From local to gcs
+    message("2. From local to GCS")
     gcs_filename <- local_to_gcs(
       x = geozip_dir,
       bucket = bucket,
       quiet = quiet
     )
+
+    message("3. From GCS to Earth Engine")
     gcs_to_ee_table(
       gs_uri = gcs_filename,
       overwrite = overwrite,
-      assetId = assetId
+      assetId = assetId,
+      command_line_tool_path = command_line_tool_path
     )
     if (isTRUE(monitoring)) {
       ee_monitoring()